+++ src/lightwin/beam_calculation/envelope_3d/specs.py
@@ -1,8 +1,13 @@
 """Define how :class:`.Envelope3D` should be configured."""
 
+from types import NoneType
+
+from lightwin.beam_calculation.deprecated_specs import (
+    apply_deprecated_flag_phi_abs,
+)
 from lightwin.beam_calculation.envelope_3d.util import ENVELOPE3D_METHODS
 from lightwin.config.key_val_conf_spec import KeyValConfSpec
-from lightwin.util.typing import EXPORT_PHASES
+from lightwin.util.typing import EXPORT_PHASES, REFERENCE_PHASE_POLICY
 
 ENVELOPE3D_CONFIG = (
     KeyValConfSpec(
@@ -32,14 +37,31 @@ ENVELOPE3D_CONFIG = (
         warning_message="Not implemented yet, will ignore this key.",
     ),
     KeyValConfSpec(
-        key="flag_phi_abs",
-        types=(bool,),
+        key="reference_phase_policy",
+        types=(str,),
         description=(
+            "Controls how cavities reference phase will be initialized."
+        ),
+        default_value="phi_0_abs",
+        allowed_values=REFERENCE_PHASE_POLICY,
+        is_mandatory=False,
+    ),
+    KeyValConfSpec(
+        key="flag_phi_abs",
+        types=(bool, NoneType),
+        description=(
+            "DEPRECATED, prefer use of `reference_phase_policy`. "
             "If the field maps phases should be absolute (no implicit "
             "rephasing after a failure)."
         ),
-        default_value=True,
+        default_value=None,
         is_mandatory=False,
+        warning_message=(
+            "The ``flag_phi_abs`` option is deprecated, prefer using the "
+            "``reference_phase_policy``.\nflag_phi_abs=False -> "
+            "reference_phase_policy='phi_0_rel'\nflag_phi_abs=True -> "
+            "reference_phase_policy='phi_0_abs'"
+        ),
     ),
     KeyValConfSpec(
         key="method",
@@ -73,4 +95,4 @@ ENVELOPE3D_CONFIG = (
     ),
 )
 
-ENVELOPE3D_MONKEY_PATCHES = {}
+ENVELOPE3D_MONKEY_PATCHES = {"_pre_treat": apply_deprecated_flag_phi_abs}