+++ tests/beam_calculator/test_envelope3d.py
@@ -6,6 +6,7 @@
 """
 
 from typing import Any
+from unittest.mock import call, patch
 
 import pytest
 from tests.pytest_helpers.simulation_output import wrap_approx
@@ -20,12 +21,27 @@ from lightwin.constants import example_config
 from lightwin.core.accelerator.accelerator import Accelerator
 from lightwin.core.accelerator.factory import NoFault
 
+# Arguments are: reference_phase_policy, n_steps_per_cell
 params = [
     pytest.param(
-        (False, 40), marks=pytest.mark.smoke, id="Envelope3D relative phase"
+        ("phi_0_rel", 40),
+        marks=pytest.mark.smoke,
+        id="Envelope3D relative phase",
     ),
     pytest.param(
-        (True, 40), marks=pytest.mark.smoke, id="Envelope3D absolute phase"
+        ("phi_0_abs", 40),
+        marks=pytest.mark.smoke,
+        id="Envelope3D absolute phase",
+    ),
+    pytest.param(
+        ("phi_s", 40),
+        marks=(pytest.mark.smoke, pytest.mark.implementation),
+        id="Envelope3D synchronous phase",
+    ),
+    pytest.param(
+        ("as_in_original_dat", 40),
+        marks=(pytest.mark.smoke, pytest.mark.implementation),
+        id="Envelope3D different phases, matching original ``DAT``",
     ),
 ]
 
@@ -37,7 +53,7 @@ def config(
 ) -> dict[str, dict[str, Any]]:
     """Set the configuration, common to all solvers."""
     out_folder = tmp_path_factory.mktemp("tmp")
-    flag_phi_abs, n_steps_per_cell = request.param
+    reference_phase_policy, n_steps_per_cell = request.param
 
     config_keys = {
         "files": "files",
@@ -50,7 +66,7 @@ def config(
         },
         "beam_calculator": {
             "tool": "Envelope3D",
-            "flag_phi_abs": flag_phi_abs,
+            "reference_phase_policy": reference_phase_policy,
             "n_steps_per_cell": n_steps_per_cell,
         },
     }
@@ -91,7 +107,12 @@ def simulation_output(
 
 @pytest.mark.envelope3d
 class TestSolver3D:
-    """Gater all the tests in a single class."""
+    """Gater all the tests in a single class.
+
+    Note that, in absence of failure, the ``reference_phase_policy`` should not
+    have any influence.
+
+    """
 
     def test_w_kin(self, simulation_output: SimulationOutput) -> None:
         """Check the beam energy at the exit of the linac."""
@@ -122,3 +143,89 @@ class TestSolver3D:
     def test_r_zdelta(self, simulation_output: SimulationOutput) -> None:
         """Verify that longitudinal transfer matrix is correct."""
         assert wrap_approx("r_zdelta", simulation_output, abs=5e-3)
+
+
+@pytest.mark.envelope3d
+@pytest.mark.implementation
+def test_deprecated_flag_phi_abs_false(
+    tmp_path_factory: pytest.TempPathFactory,
+) -> None:
+    """Check that the ``flag_phi_abs`` is considered, but warning is raised."""
+    out_folder = tmp_path_factory.mktemp("tmp")
+    config_keys = {
+        "files": "files",
+        "beam_calculator": "generic_envelope3d",
+        "beam": "beam",
+    }
+    override = {
+        "files": {"project_folder": out_folder},
+        "beam_calculator": {
+            "reference_phase_policy": "phi_0_abs",
+            "flag_phi_abs": False,
+        },
+    }
+    calls = [
+        call(
+            "Overriding ``reference_phase_policy`` following (deprecated) "
+            "flag_phi_abs = False. reference_phase_policy phi_0_abs -> "
+            "phi_0_rel"
+        ),
+        call(
+            "The ``flag_phi_abs`` option is deprecated, prefer using the "
+            "``reference_phase_policy``.\nflag_phi_abs=False -> "
+            "reference_phase_policy='phi_0_rel'\nflag_phi_abs=True -> "
+            "reference_phase_policy='phi_0_abs'"
+        ),
+    ]
+    with patch("logging.warning") as mock_warning:
+        my_config = config_manager.process_config(
+            example_config, config_keys, override=override
+        )
+        mock_warning.assert_has_calls(calls)
+        assert (
+            my_config["beam_calculator"]["reference_phase_policy"]
+            == "phi_0_rel"
+        )
+
+
+@pytest.mark.envelope3d
+@pytest.mark.implementation
+def test_deprecated_flag_phi_abs_true(
+    tmp_path_factory: pytest.TempPathFactory,
+) -> None:
+    """Check that the ``flag_phi_abs`` is considered, but warning is raised."""
+    out_folder = tmp_path_factory.mktemp("tmp")
+    config_keys = {
+        "files": "files",
+        "beam_calculator": "generic_envelope3d",
+        "beam": "beam",
+    }
+    override = {
+        "files": {"project_folder": out_folder},
+        "beam_calculator": {
+            "reference_phase_policy": "phi_s",
+            "flag_phi_abs": True,
+        },
+    }
+    calls = [
+        call(
+            "Overriding ``reference_phase_policy`` following (deprecated) "
+            "flag_phi_abs = True. reference_phase_policy phi_s -> "
+            "phi_0_abs"
+        ),
+        call(
+            "The ``flag_phi_abs`` option is deprecated, prefer using the "
+            "``reference_phase_policy``.\nflag_phi_abs=False -> "
+            "reference_phase_policy='phi_0_rel'\nflag_phi_abs=True -> "
+            "reference_phase_policy='phi_0_abs'"
+        ),
+    ]
+    with patch("logging.warning") as mock_warning:
+        my_config = config_manager.process_config(
+            example_config, config_keys, override=override
+        )
+        mock_warning.assert_has_calls(calls)
+        assert (
+            my_config["beam_calculator"]["reference_phase_policy"]
+            == "phi_0_abs"
+        )