+++ tests/beam_calculator/test_tracewin.py
@@ -10,6 +10,7 @@
 """
 
 from typing import Any
+from unittest.mock import call, patch
 
 import pytest
 from tests.pytest_helpers.simulation_output import wrap_approx
@@ -89,7 +90,12 @@ def simulation_output(
 
 @pytest.mark.tracewin
 class TestSolver3D:
-    """Gater all the tests in a single class."""
+    """Gater all the tests in a single class.
+
+    Note that, in absence of failure, the ``reference_phase_policy`` should not
+    have any influence.
+
+    """
 
     def test_w_kin(self, simulation_output: SimulationOutput) -> None:
         """Check the beam energy at the exit of the linac."""
@@ -133,3 +139,89 @@ class TestSolver3D:
     def test_r_zdelta(self, simulation_output: SimulationOutput) -> None:
         """Verify that final longitudinal transfer matrix is correct."""
         assert wrap_approx("r_zdelta", simulation_output)
+
+
+@pytest.mark.tracewin
+@pytest.mark.implementation
+def test_deprecated_flag_phi_abs_false(
+    tmp_path_factory: pytest.TempPathFactory,
+) -> None:
+    """Check that the ``flag_phi_abs`` is considered, but warning is raised."""
+    out_folder = tmp_path_factory.mktemp("tmp")
+    config_keys = {
+        "files": "files",
+        "beam_calculator": "generic_tracewin",
+        "beam": "beam",
+    }
+    override = {
+        "files": {"project_folder": out_folder},
+        "beam_calculator": {
+            "reference_phase_policy": "phi_0_abs",
+            "flag_phi_abs": False,
+        },
+    }
+    calls = [
+        call(
+            "Overriding ``reference_phase_policy`` following (deprecated) "
+            "flag_phi_abs = False. reference_phase_policy phi_0_abs -> "
+            "phi_0_rel"
+        ),
+        call(
+            "The ``flag_phi_abs`` option is deprecated, prefer using the "
+            "``reference_phase_policy``.\nflag_phi_abs=False -> "
+            "reference_phase_policy='phi_0_rel'\nflag_phi_abs=True -> "
+            "reference_phase_policy='phi_0_abs'"
+        ),
+    ]
+    with patch("logging.warning") as mock_warning:
+        my_config = config_manager.process_config(
+            example_config, config_keys, override=override
+        )
+        mock_warning.assert_has_calls(calls)
+        assert (
+            my_config["beam_calculator"]["reference_phase_policy"]
+            == "phi_0_rel"
+        )
+
+
+@pytest.mark.tracewin
+@pytest.mark.implementation
+def test_deprecated_flag_phi_abs_true(
+    tmp_path_factory: pytest.TempPathFactory,
+) -> None:
+    """Check that the ``flag_phi_abs`` is considered, but warning is raised."""
+    out_folder = tmp_path_factory.mktemp("tmp")
+    config_keys = {
+        "files": "files",
+        "beam_calculator": "generic_tracewin",
+        "beam": "beam",
+    }
+    override = {
+        "files": {"project_folder": out_folder},
+        "beam_calculator": {
+            "reference_phase_policy": "phi_s",
+            "flag_phi_abs": True,
+        },
+    }
+    calls = [
+        call(
+            "Overriding ``reference_phase_policy`` following (deprecated) "
+            "flag_phi_abs = True. reference_phase_policy phi_s -> "
+            "phi_0_abs"
+        ),
+        call(
+            "The ``flag_phi_abs`` option is deprecated, prefer using the "
+            "``reference_phase_policy``.\nflag_phi_abs=False -> "
+            "reference_phase_policy='phi_0_rel'\nflag_phi_abs=True -> "
+            "reference_phase_policy='phi_0_abs'"
+        ),
+    ]
+    with patch("logging.warning") as mock_warning:
+        my_config = config_manager.process_config(
+            example_config, config_keys, override=override
+        )
+        mock_warning.assert_has_calls(calls)
+        assert (
+            my_config["beam_calculator"]["reference_phase_policy"]
+            == "phi_0_abs"
+        )